feat: Visual parity updates, home dashboard, and compile fix
- Update design system theme (colors, typography, spacing, buttons, cards)
- Add Acme font and new string resources for calculator screens
- Rework home screen as dashboard with metric cards and pull-to-refresh
- Update all calculator screens (BMI, BMR, body fat, BP, heart rate, ideal weight, water intake)
- Add navigation from home dashboard tiles to calculator screens
- Fix exhaustive when expression in HomeUiStateHolder for OnNavigateToCalculator event
- Update Adapty subscription provider and calculators screen/state holder
- Add visual parity plan docs and update project knowledge

+++ libs/subscription/subscription-adapty/src/mobileMain/kotlin/com/cohesionbrew/healthcalculator/subscription/adapty/AdaptySubscriptionProvider.kt
@@ -224,7 +224,7 @@ internal class AdaptySubscriptionProvider : SubscriptionProvider {
 
     private fun AdaptyPrice.asGrantedAccessPrice(): Price {
         return Price(
-            amount = amount,
+            amount = amount.toFloat(),
             currencyCodeOrSymbol = currencyCode ?: currencySymbol,
             localizedString = localizedString
         )
